Remove unused default config module and document JSON configs
