+++ scripts/control_loop.py
@@ -21,7 +21,7 @@ from app.rules import RulesEngine
 from app.memory.db import Database
 from app.memory.kpis import KPICalculator
 from app.utils import setup_logging, create_alert
-from app.config.default import load_config
+from app.utils import load_config
 
 
 class ControlLoop:
